Simplify KSPDev client module
- Deprecate "repeated" methods. Repeated logs will be handled by UI in
SmartMode of the new KSPDev.Console module.
- Drop logs filtering as it will be handled by KSPDev.Console.
- Stop adding "source" since it will be calculated by
KSPDev.LogInterceptor.
