+++ Plugins/Source/KSPDev-sample-settings.cfg
@@ -1,0 +1,16 @@
+// Copy this file into GameData/KSP_Dev folder as "settings.cfg". Without any
+// settings all logging produced by KIS (or any other addon which uses KSP_Dev)
+// is silenced.
+KSPDevConfig
+{
+	name = KSPDevConfig
+	// Allowed values: NONE, ERROR, WARNING, INFO, TRACE
+	loggingLevel = INFO
+	// Allowed values: DevLogger
+	loggerName = DevLogger
+
+	DevLogger
+	{
+		// No settings for now.
+	}
+}
