--- Plugins/Source/KSP_Dev-sample-settings.cfg
@@ -1,20 +1,0 @@
-// Copy this file into GameData/KSP_Dev folder as "settings.cfg". Without any
-// settings all logging produced by KIS (or any other addon which uses KSP_Dev)
-// is silenced. When logs are enabled the game's performance and memory
-// footprint may be impacted so, don't let them be when just playing.
-KSPDevConfig
-{
-	name = KSPDevConfig
-	// Allowed values: NONE, ERROR, WARNING, INFO, TRACE
-	loggingLevel = INFO
-	// Allowed values: DevLogger
-	loggerName = DevLogger
-
-	DevLogger
-	{
-// Define which modules and/or methods should be shown in the log. Or leave
-// this section empty to allow all logs.
-//		whitelistedModule = KIS
-//		whitelistedMethod = KIS.KISAddinPickup.Internal_Update
-	}
-}
